requirements.txt
aiohttp
requests
loguru
pipecat-ai[silero,daily,openai,deepgram]
higgungface_hub
vllm
google-auth
python-dotenv